[FEAT] 폴더 구분
