+++ server/api_server/go.mod
@@ -1,0 +1,16 @@
+module github.com/Chocobone/GDG_backend_sideProject/server/api_server
+
+go 1.23.1
+
+require (
+	github.com/Chocobone/GDG_backend_sideProject/server v0.0.0-20241204113345-52c0fc98a057
+	google.golang.org/grpc v1.68.0
+)
+
+require (
+	golang.org/x/net v0.29.0 // indirect
+	golang.org/x/sys v0.25.0 // indirect
+	golang.org/x/text v0.18.0 // indirect
+	google.golang.org/genproto/googleapis/rpc v0.0.0-20240903143218-8af14fe29dc1 // indirect
+	google.golang.org/protobuf v1.35.2 // indirect
+)
